feat(site): Eleventy高速化のため事前処理コマンドを追加
