+++ .github/workflows/ci.yml
@@ -101,6 +101,8 @@ jobs:
     - run: npm ci
     - name: Generate feed
       run: npm run feed-generate
+    - name: Prepare Site
+      run: npm run site-prepare
     - name: Generate site
       run: npm run site-build
     - name: Archive feed
