+++ .github/workflows/generate-feed.yml
@@ -48,6 +48,9 @@ jobs:
       - name: Generate feed
         run: npm run feed-generate
 
+      - name: Prepare Site
+        run: npm run site-prepare
+
       - name: Generate site
         run: npm run site-build
