+++ src/cli/generate-feed-command.ts
@@ -1,21 +1,15 @@
 import * as path from 'node:path';
 import * as url from 'node:url';
+import constants from '../common/constants';
+import { FeedCrawler } from '../feed/utils/feed-crawler';
+import { FeedGenerator } from '../feed/utils/feed-generator';
+import { FeedStorer } from '../feed/utils/feed-storer';
+import { FeedValidator } from '../feed/utils/feed-validator';
+import { logger } from '../feed/utils/logger';
 import { FEED_INFO_LIST } from '../resources/feed-info-list';
-import { FeedCrawler } from './utils/feed-crawler';
-import { FeedGenerator } from './utils/feed-generator';
-import { ImagePrecacher as FeedImagePrecacher } from './utils/feed-image-precacher';
-import { FeedStorer } from './utils/feed-storer';
-import { FeedValidator } from './utils/feed-validator';
-import { logger } from './utils/logger';
 
 const dirName = url.fileURLToPath(new URL('.', import.meta.url));
 
-const FEED_FETCH_CONCURRENCY = 50;
-const FEED_OG_FETCH_CONCURRENCY = 20;
-const FETCH_IMAGE_CONCURRENCY = 100;
-const FILTER_ARTICLE_DATE = new Date(Date.now() - 8 * 24 * 60 * 60 * 1000);
-const MAX_FEED_DESCRIPTION_LENGTH = 200;
-const MAX_FEED_CONTENT_LENGTH = 500;
 const STORE_FEEDS_DIR_PATH = path.join(dirName, '../site/feeds');
 const STORE_BLOG_FEEDS_DIR_PATH = path.join(dirName, '../site/blog-feeds');
 
@@ -23,15 +17,14 @@ const feedCrawler = new FeedCrawler();
 const feedGenerator = new FeedGenerator();
 const feedValidator = new FeedValidator();
 const feedStorer = new FeedStorer();
-const feedImagePrecacher = new FeedImagePrecacher();
 
 (async () => {
   // フィード取得
   const crawlFeedsResult = await feedCrawler.crawlFeeds(
     FEED_INFO_LIST,
-    FEED_FETCH_CONCURRENCY,
-    FEED_OG_FETCH_CONCURRENCY,
-    FILTER_ARTICLE_DATE,
+    constants.feedFetchConcurrency,
+    constants.feedOgFetchConcurrency,
+    new Date(Date.now() - constants.aggregateFeedDurationInHours * 60 * 60 * 1000),
   );
 
   // まとめフィード作成
@@ -40,8 +33,8 @@ const feedImagePrecacher = new FeedImagePrecacher();
     crawlFeedsResult.feedItems,
     ogObjectMap,
     crawlFeedsResult.feedItemHatenaCountMap,
-    MAX_FEED_DESCRIPTION_LENGTH,
-    MAX_FEED_CONTENT_LENGTH,
+    constants.maxFeedDescriptionLength,
+    constants.maxFeedContentLength,
   );
 
   // ファイル出力
@@ -56,22 +49,6 @@ const feedImagePrecacher = new FeedImagePrecacher();
     );
   } catch (e) {
     const error = new Error('Failed to store feeds', {
-      cause: e,
-    });
-    console.error(error);
-    throw error;
-  }
-
-  // 画像の事前キャッシュ
-  try {
-    await feedImagePrecacher.fetchAndCacheFeedImages(
-      crawlFeedsResult.feeds,
-      crawlFeedsResult.feedItems,
-      ogObjectMap,
-      FETCH_IMAGE_CONCURRENCY,
-    );
-  } catch (e) {
-    const error = new Error('Failed to cache feed images', {
       cause: e,
     });
     console.error(error);
