+++ src/common/constants.ts
@@ -37,4 +37,13 @@ export default {
   // サイトの追加方法のリンク
   howToAddSiteLink:
     'https://github.com/yamadashy/tech-blog-rss-feed#%E3%82%B5%E3%82%A4%E3%83%88%E3%81%AE%E8%BF%BD%E5%8A%A0%E6%96%B9%E6%B3%95',
+
+  // 処理の設定
+  feedFetchConcurrency: 50, // フィードを取得する並列数
+  feedOgFetchConcurrency: 20, // OG情報を取得する並列数
+  aggregateFeedDurationInHours: 8 * 24, // まとめフィードの対象となる時間の範囲
+  maxFeedDescriptionLength: 200, // フィードのdescriptionの最大文字数
+  maxFeedContentLength: 500, // フィードのcontentの最大文字数
+  processImageConcurrency: 50, // 画像の処理の並列数。画像取得と変換
+  eleventyFetchConcurrency: 50, // Eleventyの画像取得の並列数
 };
