+++ tsconfig.json
@@ -11,7 +11,8 @@
     "noImplicitAny": true,
     "removeComments": true,
     "skipLibCheck": true,
-    "sourceMap": true
+    "sourceMap": true,
+    "resolveJsonModule": true
   },
   "include": ["src/**/*", "tests/**/*", "vitest.config.ts", "eleventy.config.ts"]
 }
